misinformation-radar/backend/requirements.txt
fastapi
uvicorn[standard]
spacy
psycopg2-binary
pydantic

# ML dependencies (placeholders for now)
torch
transformers
scikit-learn
numpy
pandas
networkx
